services/document-chunking/requirements.txt
# Document Chunking Service - Dependencies
# Splits PDF/DOCX/TXT files into semantic text chunks

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Document processing
langchain>=0.1.0
langchain-text-splitters>=0.0.1
pypdf>=3.17.1
python-docx>=1.1.0

# AWS S3 integration
boto3>=1.33.0

# Utilities
python-multipart>=0.0.6
pydantic>=2.5.0

requests==2.31.0
